Added Lombok and Gson
Removed json-io in favour of Gson and added Lombok while I was at it.
Dependencies no longer load from dependencies.gradle. They are now located in Config.glassDeps.
This commit is very unstable (most notably options not opening and if there are no instances, a null instance will be made)