Add session saving, add MS login support, add new logging system for minecraft
This is a very experimental commit, and is questionable in its reliability.
Add session saving.
Add Microsoft login support.
Add a new logging system for minecraft that saves in `.glass-launcher/glass-logs/minecraft`.
Actually add `-h` as a parameter.
Fixed a rare password box bug where it would censor the hint.
Added theme toggling, for those who dislike the new theme.

The login verifier for creating an instance does not support MS logins or sessions yet, as that has yet to be changed to use the new system.